==> 03_Scripts/02_Dynamic_time_warping_distance.ipynb <==
import os
import random
import ast
import numpy as np
import pandas as pd # use pandas for more functionality
from dtaidistance import dtw
from dask import delayed, compute
from dask.distributed import Client, default_client
# Close existing client if any
try:
    client = default_client()
    client.close()
except ValueError:
    pass
client = Client() # Start a new Dask client before importing modin pandas
# import modin.pandas as pd # use modin to speed things up (dont use modin wiht Dask)

# get working directory as paerent directory of current directory
cwd = os.getcwd()
pwd = os.path.dirname(cwd)

def create_binary_stream(row):
    # Unpack indices and stream length from the row
    indices, stream_length = row['event_timestamps'], row['event_length']

    # Initialize a NumPy array of zeros
    binary_stream = np.zeros(int(stream_length), dtype=int)

    if len(indices) != 0:

        # Ensure indices are integers
        indices = [int(i) for i in indices if isinstance(i, (int, float)) and not np.isnan(i)]

        # Convert indices to a NumPy array and filter out-of-bound indices
        indices = np.array(indices)
        valid_indices = indices[(0 <= indices) & (indices < int(stream_length))]

        # Set the specified indices to 1
        binary_stream[valid_indices] = 1

    return binary_stream

data_df = pd.read_csv(pwd + "/02_Clean_data/01_recording_event_times_labels_binary.csv")
# Convert string representations of lists to actual lists
data_df['event_timestamps'] = data_df['event_timestamps'].apply(ast.literal_eval)
# get the binary stream for each row
data_df['binary_stream'] = data_df.apply(create_binary_stream, axis=1)

data_df

# dtw.distance(data_df['binary_stream'].iloc[0], data_df['binary_stream'].iloc[1])

# dtw.distance(data_df['event_timestamps'].iloc[0], data_df['event_timestamps'].iloc[1])

# Distance/Similarity Measurement of data

data_df_subset = data_df.iloc[0:5]

data_df_subset

@delayed
def calculate_dtw(i, j):
    distance = dtw.distance(data_df_subset['event_timestamps'].iloc[i], data_df_subset['event_timestamps'].iloc[j])
    return i, j, distance

pairs = [(i, j) for i in range(len(data_df_subset)) for j in range(i + 1, len(data_df_subset))]

delayed_results = [calculate_dtw(i, j) for i, j in pairs]

results = compute(*delayed_results)

dtw_matrix = np.full((len(data_df_subset), len(data_df_subset)), None, dtype=float)

for i, j, distance in results:
    dtw_matrix[i][j] = distance
    dtw_matrix[j][i] = distance

np.fill_diagonal(dtw_matrix, 0)

# Remember to close the client when done
client.close()

# dtw_matrix now contains all the pairwise DTW distances
pd.DataFrame(dtw_matrix)

# Function to generate a sparse binary time series and its indices
def generate_sparse_binary_time_series(min_length=10, max_length=20):
    # Randomly choose a length for the time series
    length = random.randint(min_length, max_length)
    # Generate the time series with random 0s and 1s
    time_series = [random.choice([0, 1]) for _ in range(length)]
    # Find the indices where the value is 1
    indices = [index for index, value in enumerate(time_series) if value == 1]
    return time_series, indices

# Create the DataFrame
# Here, we generate 5 rows as an example. You can generate more or fewer rows as needed.
data = [generate_sparse_binary_time_series() for _ in range(5)]
df = pd.DataFrame(data, columns=['binary_timeseries', 'event_timestamps'])

# rewrite testing df
data_df_subset = df

client = Client()

@delayed
def calculate_dtw(i, j):
    distance = dtw.distance(data_df_subset['event_timestamps'].iloc[i], data_df_subset['event_timestamps'].iloc[j])
    return i, j, distance

pairs = [(i, j) for i in range(len(data_df_subset)) for j in range(i + 1, len(data_df_subset))]

delayed_results = [calculate_dtw(i, j) for i, j in pairs]

results = compute(*delayed_results)

dtw_matrix = np.full((len(data_df_subset), len(data_df_subset)), None, dtype=float)

for i, j, distance in results:
    dtw_matrix[i][j] = distance
    dtw_matrix[j][i] = distance

np.fill_diagonal(dtw_matrix, 0)

# Remember to close the client when done
client.close()

# dtw_matrix now contains all the pairwise DTW distances
pd.DataFrame(dtw_matrix)

client = Client()

@delayed
def calculate_dtw(i, j):
    distance = dtw.distance(data_df_subset['binary_timeseries'].iloc[i], data_df_subset['binary_timeseries'].iloc[j])
    return i, j, distance

pairs = [(i, j) for i in range(len(data_df_subset)) for j in range(i + 1, len(data_df_subset))]

delayed_results = [calculate_dtw(i, j) for i, j in pairs]

results = compute(*delayed_results)

dtw_matrix = np.full((len(data_df_subset), len(data_df_subset)), None, dtype=float)

for i, j, distance in results:
    dtw_matrix[i][j] = distance
    dtw_matrix[j][i] = distance

np.fill_diagonal(dtw_matrix, 0)

# Remember to close the client when done
client.close()

# dtw_matrix now contains all the pairwise DTW distances
pd.DataFrame(dtw_matrix)

import seaborn as sns
import matplotlib.pyplot as plt
import plotly.graph_objects as go

fig = go.Figure()
for i in range(len(data_df_subset)):
    ts = data_df_subset['binary_timeseries'].iloc[i]

    # Plotting using plotly
    fig.add_trace(go.Scatter(x=list(range(len(ts))), y=ts, mode='lines'))

    # Plotting using seaborn
    # sns.lineplot(data=ts, markers=True, dashes=False)
# plt.show()
fig.show()
